Restructing code
