+++ .flake8
@@ -2,5 +2,5 @@
 ignore = E501,W503
 builtins = unicode
 tee = True
-exclude = env,venv,examples,*_pb2*
+exclude = env,venv,examples,*_pb2*,__init__.py
 max-complexity = 9
