+++ docker-compose.yml
@@ -24,7 +24,7 @@ x-test-build: &test-build
   image: "needlestack-test"
 
 x-environment: &environment
-  environment: 
+  environment:
     PYTHONPATH: '/app'
     LANG: 'en_US.UTF-8'
 
@@ -42,7 +42,7 @@ x-searcher-grpc-template: &searcher-grpc-template
     - examples/run_searcher.py
 
 x-zookeeper-image: &zookeeper-image
-  image: zookeeper:3.5.5
+  image: zookeeper:3.5.6
 
 
 #######################
@@ -93,7 +93,7 @@ services:
       - 2181:2181
     environment:
       ZOO_MY_ID: 1
-      ZOO_SERVERS: server.1=0.0.0.0:2888:3888 server.2=zoo2:2888:3888 server.3=zoo3:2888:3888
+      ZOO_SERVERS: server.1=0.0.0.0:2888:3888;2181 server.2=zoo2:2888:3888;2181 server.3=zoo3:2888:3888;2181
 
   zoo2:
     <<: *zookeeper-image
@@ -102,7 +102,7 @@ services:
       - 2182:2181
     environment:
       ZOO_MY_ID: 2
-      ZOO_SERVERS: server.1=zoo1:2888:3888 server.2=0.0.0.0:2888:3888 server.3=zoo3:2888:3888
+      ZOO_SERVERS: server.1=zoo1:2888:3888;2181 server.2=0.0.0.0:2888:3888;2181 server.3=zoo3:2888:3888;2181
 
   zoo3:
     <<: *zookeeper-image
@@ -111,7 +111,7 @@ services:
       - 2183:2181
     environment:
       ZOO_MY_ID: 3
-      ZOO_SERVERS: server.1=zoo1:2888:3888 server.2=zoo2:2888:3888 server.3=0.0.0.0:2888:3888
+      ZOO_SERVERS: server.1=zoo1:2888:3888;2181 server.2=zoo2:2888:3888;2181 server.3=0.0.0.0:2888:3888;2181
 
   test:
     <<: *test-build
