+++ docs/needlestack.apis.rst
@@ -60,18 +60,18 @@ needlestack.apis.health\_pb2\_grpc module
    :undoc-members:
    :show-inheritance:
 
-needlestack.apis.neighbors\_pb2 module
---------------------------------------
+needlestack.apis.indices\_pb2 module
+------------------------------------
 
-.. automodule:: needlestack.apis.neighbors_pb2
+.. automodule:: needlestack.apis.indices_pb2
    :members:
    :undoc-members:
    :show-inheritance:
 
-needlestack.apis.neighbors\_pb2\_grpc module
---------------------------------------------
+needlestack.apis.indices\_pb2\_grpc module
+------------------------------------------
 
-.. automodule:: needlestack.apis.neighbors_pb2_grpc
+.. automodule:: needlestack.apis.indices_pb2_grpc
    :members:
    :undoc-members:
    :show-inheritance:
