+++ docs/needlestack.cluster_managers.rst
@@ -4,6 +4,14 @@ needlestack.cluster\_managers package
 Submodules
 ----------
 
+needlestack.cluster\_managers.manager module
+--------------------------------------------
+
+.. automodule:: needlestack.cluster_managers.manager
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 needlestack.cluster\_managers.zookeeper module
 ----------------------------------------------
