+++ docs/needlestack.indices.rst
@@ -1,0 +1,30 @@
+needlestack.indices package
+===========================
+
+Submodules
+----------
+
+needlestack.indices.faiss\_indices module
+-----------------------------------------
+
+.. automodule:: needlestack.indices.faiss_indices
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+needlestack.indices.index module
+--------------------------------
+
+.. automodule:: needlestack.indices.index
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: needlestack.indices
+   :members:
+   :undoc-members:
+   :show-inheritance:
