--- docs/needlestack.neighbors.rst
@@ -1,22 +1,0 @@
-needlestack.neighbors package
-=============================
-
-Submodules
-----------
-
-needlestack.neighbors.faiss\_indices module
--------------------------------------------
-
-.. automodule:: needlestack.neighbors.faiss_indices
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: needlestack.neighbors
-   :members:
-   :undoc-members:
-   :show-inheritance:
