+++ needlestack/collections/shard.py
@@ -1,10 +1,10 @@
-from typing import Tuple, List
+from typing import List
 
 import numpy as np
 
-from needlestack.apis import neighbors_pb2
+from needlestack.apis import indices_pb2
 from needlestack.apis import collections_pb2
-from needlestack.neighbors import SpatialIndex
+from needlestack.indices import BaseIndex
 
 
 class Shard(object):
@@ -14,13 +14,13 @@ class Shard(object):
     Attributes:
         name: Name of shard
         weight: Weight of shard
-        index: SpatialIndex for kNN queries
+        index: BaseIndex for kNN queries
         enable_id_to_vector: Enable retrieving vector from id
     """
 
     name: str
     weight: float
-    index: SpatialIndex
+    index: BaseIndex
     enable_id_to_vector: bool = False
 
     @classmethod
@@ -32,18 +32,14 @@ class Shard(object):
     def populate_from_proto(self, proto: collections_pb2.Shard):
         self.name = proto.name
         self.weight = proto.weight
-        self.index = SpatialIndex.from_proto(proto.index)
+        self.index = BaseIndex.from_proto(proto.index)
 
     def load(self):
         self.index.enable_id_to_vector = self.enable_id_to_vector
         self.index.load()
 
-    def query(
-        self, X: np.ndarray, k: int
-    ) -> List[Tuple[float, neighbors_pb2.Metadata]]:
+    def query(self, X: np.ndarray, k: int) -> List[indices_pb2.SearchResultItem]:
         return self.index.query(X, k)[0]
 
-    def get_vector_and_metadata(
-        self, id: str
-    ) -> Tuple[np.ndarray, neighbors_pb2.Metadata]:
-        return self.index.get_vector_and_metadata(id)
+    def retrieve(self, id: str) -> indices_pb2.RetrievalResultItem:
+        return self.index.retrieve(id)
