+++ needlestack/indices/__init__.py
@@ -1,0 +1,1 @@
+from needlestack.indices.index import BaseIndex
