+++ needlestack/servicers/factory.py
@@ -40,7 +40,7 @@ def create_zookeeper_client(config: BaseConfig) -> ZookeeperClusterManager:
         config: Config with settings on how to set up a Zookeeper client
     """
     return ZookeeperClusterManager(
-        config.CLUSTER_NAME, config.HOSTPORT, config.ZOOKEEPER_HOSTS
+        config.CLUSTER_NAME, config.hostport, config.ZOOKEEPER_HOSTS
     )
