+++ pytest.ini
@@ -1,4 +1,4 @@
 
 [pytest]
-addopts = --verbose --cov=needlestack --cov-report=term --cov-report=html:htmlcov --capture=no
+addopts = --cov=needlestack --cov-report=term --cov-report=html:htmlcov --capture=no
 python_files = tests/*.py
